rename files to match modules
